Griddledogs!

+++ modular/Neu_Food/code/cooked/cooked_meat.dm
@@ -372,6 +372,7 @@
 	icon_state = "wiener"
 	faretype = FARE_NEUTRAL
 	fried_type = null
+	tastes = list("tender wiener" = 1)
 	bonus_reagents = list(/datum/reagent/consumable/nutriment = SNACK_DECENT)
 	rotprocess = SHELFLIFE_EXTREME
 
@@ -384,6 +385,14 @@
 			if(do_after(user,short_cooktime, target = src))
 				add_sleep_experience(user, /datum/skill/craft/cooking, user.STAINT)
 				new /obj/item/reagent_containers/food/snacks/rogue/friedegg/sausage(loc)
+				qdel(I)
+				qdel(src)
+	if(istype(I, /obj/item/grown/log/tree/stick))
+		if(isturf(loc)&& (found_table))
+			playsound(get_turf(user), 'sound/foley/dropsound/gen_drop.ogg', 30, TRUE, -1)
+			if(do_after(user,short_cooktime, target = src))
+				add_sleep_experience(user, /datum/skill/craft/cooking, user.STAINT)
+				new /obj/item/reagent_containers/food/snacks/rogue/wienerstick(loc)
 				qdel(I)
 				qdel(src)
 	else
